LPS-152834 Moves to item.selector package
