+++ modules/apps/info/info-collection-provider-item-selector-web/src/main/java/com/liferay/info/collection/provider/item/selector/web/internal/item/selector/RelatedInfoItemCollectionProviderItemDescriptor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.info.collection.provider.item.selector.web.internal;
+package com.liferay.info.collection.provider.item.selector.web.internal.item.selector;
 
 import com.liferay.frontend.taglib.clay.servlet.taglib.VerticalCard;
 import com.liferay.info.collection.provider.RelatedInfoItemCollectionProvider;
